+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2 - 2024/09/28
+
+* Changed the `initialSeconds` parameter to `initialDuration`.
+    * Changed the format from double to a `Duration()`
+* Added a `timeFormatter` parameter to format the time.
+    * Added an example time formatter
+
 ## 0.0.1 - 2024/09/25
 
 * Initial Release